requirements.txt
python-dotenv
flask
flask-wtf
flask-sqlalchemy
flask-migrate
flask-login
email-validator
requests
elasticsearch
openai